gRPC/server-2 (Algorithms)/TB/requirements.txt
numpy
sympy
protobuf==3.20.*
grpcio
requests
certifi
opencv-python
opencv-contrib-python
tritonclient
confluent-kafka==1.8.2